stream and declension appeared
